+++ ErrorCode/Domain/TestClass.cs
@@ -27,7 +27,7 @@ namespace ErrorCode.Domain
                          .ToArray();
         }
 
-        public IReadOnlyList<TestState> Run(double interval = Constants.DefaultInterval)
+        public IReadOnlyList<TestState> Run(int interval = Constants.DefaultInterval)
         {
             object instance = CreateTestInstance();
